move FeedbackTitle component to its own file
